# Improve .NET 8 stream decompression perf